app/requirements.txt
FastAPI
uvicorn
torch
torchvision
Pillow
numpy
torchmetrics
python-multipart